feat/full kafka workflow testing
