+++ tests/test_full_kafka_flow.py
@@ -1,0 +1,106 @@
+import pytest
+import asyncio
+import random
+import string
+import time
+from aiokafka import AIOKafkaProducer
+from scidx.client.kafka_client import KafkaClient
+from scidx.client.get_kafka_data import consume_kafka_messages
+
+# Constants
+API_URL = "http://localhost:8000"
+KAFKA_HOST = '155.101.6.194'
+KAFKA_PORT = '9092'
+KAFKA_TOPIC_PREFIX = 'random_topic_example_'
+OWNER_ORG = "test_org5"
+
+# Function to generate a unique Kafka topic name
+def generate_unique_kafka_topic():
+    return KAFKA_TOPIC_PREFIX + ''.join(random.choices(string.ascii_lowercase + string.digits, k=6))
+
+class KafkaProducer:
+    def __init__(self, host, port, topic):
+        self.host = host
+        self.port = port
+        self.topic = topic
+
+    async def send_random_messages(self, num_messages):
+        producer = AIOKafkaProducer(bootstrap_servers=f"{self.host}:{self.port}")
+        await producer.start()
+        try:
+            for _ in range(num_messages):
+                message = ''.join(random.choices(string.ascii_letters + string.digits, k=20))
+                await producer.send_and_wait(self.topic, message.encode('utf-8'))
+                print(f"Sent message: {message}")
+        finally:
+            await producer.stop()
+
+@pytest.fixture
+def kafka_client():
+    return KafkaClient(API_URL)
+
+@pytest.mark.asyncio
+async def test_kafka_full_flow(kafka_client):
+    # Step 1: Generate a unique Kafka topic name
+    kafka_topic = generate_unique_kafka_topic()
+
+    # Step 2: Send random messages to Kafka topic
+    producer = KafkaProducer(KAFKA_HOST, KAFKA_PORT, kafka_topic)
+    await producer.send_random_messages(num_messages=10)
+
+    # Step 3: Register Kafka topic as dataset
+    dataset_data = {
+        "dataset_name": kafka_topic,
+        "dataset_title": "Random Topic Example",
+        "owner_org": OWNER_ORG,
+        "kafka_topic": kafka_topic,
+        "kafka_host": KAFKA_HOST,
+        "kafka_port": KAFKA_PORT,
+        "dataset_description": "This is a randomly generated Kafka topic registered as a CKAN dataset."
+    }
+    kafka_client.register_kafka(**dataset_data)
+
+    # Add a delay to ensure the dataset is indexed
+    time.sleep(5)
+
+    # Step 4: Retrieve Kafka dataset information from API
+    kafka_datasets = kafka_client.search_kafka(kafka_topic=kafka_topic, kafka_host=KAFKA_HOST, kafka_port=KAFKA_PORT)
+
+    # Ensure that at least one dataset is returned
+    assert len(kafka_datasets) > 0, "No Kafka datasets found"
+
+    dataset_info = next((ds for ds in kafka_datasets if ds['resources'][0]['kafka_topic'] == kafka_topic), None)
+    assert dataset_info is not None, f"Kafka topic {kafka_topic} not found in datasets"
+
+    # Step 5: Consume messages from Kafka topic using retrieved information
+    resource = dataset_info['resources'][0]
+    kafka_host = resource['kafka_host']
+    kafka_port = resource['kafka_port']
+    kafka_topic = resource['kafka_topic']
+
+    print(f"Starting to consume messages from topic: {kafka_topic}")
+    messages_task = asyncio.create_task(consume_kafka_messages(
+        topic=kafka_topic,
+        host=kafka_host,
+        port=kafka_port
+    ))
+
+    try:
+        await asyncio.sleep(10)
+    except asyncio.CancelledError:
+        pass
+    finally:
+        messages_task.cancel()
+
+    try:
+        messages = await messages_task
+    except asyncio.CancelledError:
+        messages = await messages_task
+
+    assert len(messages) > 0, f"No messages received from topic {kafka_topic}"
+    print(f"Messages from topic {kafka_topic}:\n")
+    for message in messages:
+        print(message)
+
+if __name__ == "__main__":
+    pytest.main([__file__])
